Make builder.NewController public
This allows a user to have the builder create custom controller types.  

Also add an ObjectType() method on Kind sources so that the objects type can accessed after the source is created.

These changes allow a user implementing a controller to observe all the Kube client watches defined, which can be helpful to then use that information to check to see a reconcile is accessing resources which are not part of the watch set for that controller.

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>